--- src/main/java/lotto/LottoMain.java
@@ -1,7 +1,0 @@
-package lotto;
-
-public class LottoMain {
-    public static void main(String[] args) {
-
-    }
-}
