+++ src/main/java/lotto/controller/LottoMain.java
@@ -1,0 +1,17 @@
+package lotto.controller;
+
+import lotto.domain.WinningLottoNumbers;
+
+public class LottoMain {
+    public static void main(String[] args) {
+//        UserPurchase userPurchase = InputView.getUserPurchase();
+//
+//        LottoMachine lottoMachine = new LottoMachine();
+//
+//        PurchasedLottoTickets purchasedLottoTickets = lottoMachine.purchaseLottoTicket(userPurchase);
+//        OutputView.printPurchasedLottoTickets(purchasedLottoTickets);
+//
+//        WinningLottoNumbers winningLottoNumbers = InputView.getWinningLottoNumbers();
+//        OutputView(lottoMachine.getLottoResult(purchasedLottoTickets, winningLottoNumbers));
+    }
+}
