+++ src/test/java/lotto/domain/LottoGeneratorTest.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.Arrays;
 import org.assertj.core.api.Assertions;
@@ -20,32 +19,17 @@ public class LottoGeneratorTest {
         )
     );
 
-    @DisplayName("구입 금액이 1000원 단위일 시 정상")
-    @Test
-    void Should_Not_ThrowException_When_PurchasePriceExactlyDividedByThousand() {
-        assertThatCode(() -> new LottoGenerator(1000))
-            .doesNotThrowAnyException();
-    }
-
-    @DisplayName("구입 금액이 1000원 단위가 아닐 시 에러")
-    @Test
-    void Should_ThrowException_When_PurchasePriceNotDividedByThousand() {
-        Assertions.assertThatThrownBy(() ->
-            new LottoGenerator(1200)
-        ).isInstanceOf(IllegalArgumentException.class);
-    }
-
     @DisplayName("구입 개수만큼의 로또 생성")
     @Test
     void Should_EqualToExpected_When_Purchased() {
-        LottoGenerator lottoGenerator = new LottoGenerator(14000);
-        Assertions.assertThat(lottoGenerator.generatePurchasedTickets().size()).isEqualTo(14);
+        LottoGenerator lottoGenerator = new LottoGenerator();
+        Assertions.assertThat(lottoGenerator.generatePurchasedTickets(14).size()).isEqualTo(14);
     }
 
     @DisplayName("당첨 번호와 보너스 번호 생성")
     @Test
     void Should_Return_WinningLottoNumbers_When_WinnerNumberDecided() {
-        LottoGenerator lottoGenerator = new LottoGenerator(14000);
+        LottoGenerator lottoGenerator = new LottoGenerator();
         LottoNumber bonusNumber = new LottoNumber(7);
 
         WinningLottoNumbers winningLottoNumbers = lottoGenerator
