+++ todo.md
@@ -1,6 +1,10 @@
 # todo
-- LottoNumber implements Comparable 한 다음에 나중에 LottoTicket에서 오름차순으로 sort하기
+- LottoNumber implements Comparable 한 다음에 LottoTicket에서 오름차순으로 sort하기
 - LottoMatchType 관련 함수 LottoMatchType 클래스 내부로 이동.
+- LottoGenerator new ArrayList()<>; 생성자 내부에서 초기화.
+- LottoGenerator.generatePurchasedTickets() 에서 List<LottoTicket> -> PurchasedLottoTickets로 변경.
+- LottoComparator 함수 수정.
+- LottoGenerator 생성자 수정.
 
 # done
 - LottoTicket 에서 allNumbers, shuffle 분리. 일급컬렉션으로.
